Test commit

+++ app/src/main/java/com/example/bd_avtorizatsia/DBHelper.java
@@ -16,10 +16,9 @@ public class DBHelper extends SQLiteOpenHelper {
     public static final String KEY_PASSWORD = "password";
 
     public static final String KEY_ID = "_id";
-    public static final String KEY_FAMILIA = "familia";
-    public static final String KEY_NAME = "name";
     public static final String KEY_TOVAR = "tovar";
-    public static final String KEY_MAIL = "mail";
+    public static final String KEY_KOLVO = "kolvo";
+    public static final String KEY_PRICE = "price";
 
     public DBHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -27,16 +26,23 @@ public class DBHelper extends SQLiteOpenHelper {
 
     @Override
     public void onCreate(SQLiteDatabase db) {
-        db.execSQL("create table " + TABLE_CONTACTS + "(" + KEY_ID
-                + " integer primary key," + KEY_FAMILIA + " text," + KEY_NAME + " text," + KEY_TOVAR + " text," + KEY_MAIL + " text" + ")");
+
         db.execSQL("create table " + TABLE_USERS + "(" + KEY_IDN + " integer primary key," + KEY_USER + " text," + KEY_PASSWORD + " text" + ")");
+
+        db.execSQL("create table " + TABLE_CONTACTS + "(" + KEY_ID
+                + " integer primary key,"  + KEY_TOVAR + " text," + KEY_KOLVO + " text," +  KEY_PRICE + " text" + ")");
+
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("drop table if exists " + TABLE_CONTACTS);
-        db.execSQL("drop table if exists " + TABLE_USERS);
+        if(oldVersion > newVersion){
+            db.execSQL("drop table if exists " + TABLE_CONTACTS);
+            onCreate(db);
 
-        onCreate(db);
+            db.execSQL("drop table if exists " + TABLE_USERS);
+            onCreate(db);
+        }
+
     }
 }
